refactor: :recycle: move get_settings function to api_settings module
