+++ app/api/api_v1/endpoints/characters.py
@@ -14,7 +14,7 @@ from app.api.api_v1.dependencies import (
 from app.api.api_v1.dependencies.filter_param_matcher import filter_param_matcher
 from app.api.api_v1.endpoints.util import get_character_details
 from app.api.api_v1.pagination import paginate_search_results
-from app.config import get_settings
+from app.config.api_settings import get_settings
 from app.data.cache import cached_data
 from app.data.encoding import get_codepoint_string
 from app.docs.dependencies.custom_parameters import (
