+++ app/api/api_v1/endpoints/planes.py
@@ -3,7 +3,7 @@ from typing import Annotated
 from fastapi import APIRouter, Path
 
 import app.db.models as db
-from app.config import get_settings
+from app.config.api_settings import get_settings
 from app.data.cache import cached_data
 
 router = APIRouter()
