+++ app/docs/api_docs/content/intro.py
@@ -1,9 +1,11 @@
 # flake8: noqa
-from app.config import get_settings
+from app.config.api_settings import get_settings
 from app.core.util import get_unicode_version_release_date
 
+settings = get_settings()
+
 INTRODUCTION = f"""
-<p>This API provides access to detailed information for all characters, blocks and planes in <a href="https://www.unicode.org/versions/Unicode{get_settings().UNICODE_VERSION}/" rel="noopener noreferrer" target="_blank">version {get_settings().UNICODE_VERSION} of the Unicode Standard</a> (released {get_unicode_version_release_date(get_settings().UNICODE_VERSION)}). In an attempt to adhere to the tenants of <a href="http://en.wikipedia.org/wiki/Representational_State_Transfer" rel="noopener noreferrer" target="_blank">REST</a>, the API is organized around the following principles:</p>
+<p>This API provides access to detailed information for all characters, blocks and planes in <a href="https://www.unicode.org/versions/Unicode{settings.UNICODE_VERSION}/" rel="noopener noreferrer" target="_blank">version {settings.UNICODE_VERSION} of the Unicode Standard</a> (released {get_unicode_version_release_date(settings.UNICODE_VERSION)}). In an attempt to adhere to the tenants of <a href="http://en.wikipedia.org/wiki/Representational_State_Transfer" rel="noopener noreferrer" target="_blank">REST</a>, the API is organized around the following principles:</p>
 <ul class="api-principles">
     <li>URLs are predictable and resource-oriented.</li>
     <li>Uses standard HTTP verbs and response codes.</li>
